refactor -> add LeftScrollContainer which has VBox,ScrollPane,ScrollBar
